Add files via upload
These are updates to HOX and XNO3 heterogeneous reactions that make the reactions of different halogens (Cl, Br, and I) for these species consistent with one another. For reactions of iodine where we don't know the rate constants, I simply use the same rate constant as for bromine. This should be a lower limit for the iodine reaction rates.